thaw_fingerprintd: Use strncmp() instead of memcmp()
memcmp() requires specific filename, and the fingerprintd hal can be upgraded to a newer version which will require us to modify the code again, hence let's switch to strncmp().

This makes the code more efficient in a way as we're using sizeof() which is determined at compile-time, so the compiler will calculate the size of the string and make it a constant.

Suggested-by: Justin Crawford <Justin@stacksmash.net>
Change-Id: I65fdcb94e372472adcebb0d128d6a769132b1700
Signed-off-by: Justin Crawford <Justin@stacksmash.net>
Signed-off-by: Cyber Knight <cyberknight755@gmail.com>
[Jprimero15: also some gx changes while at it]
Signed-off-by: Jprimero15 <jprimero15@aospa.co>